* Add release date in release notes. * Regenerate existing changelog to include release dates. * Minor class move.
