+++ daikon-documentation/src/main/java/org/talend/daikon/finders/ReleaseDateFinder.java
@@ -1,0 +1,11 @@
+package org.talend.daikon.finders;
+
+import java.util.Date;
+
+public interface ReleaseDateFinder {
+
+    /**
+     * @return The {@link Date} of the release.
+     */
+    Date find();
+}
