+++ daikon-documentation/src/main/java/org/talend/daikon/finders/git/MiscGitItemFinder.java
@@ -1,10 +1,11 @@
-package org.talend.daikon.finders;
+package org.talend.daikon.finders.git;
 
 import java.util.regex.Matcher;
 import java.util.stream.Stream;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import org.talend.daikon.finders.ItemFinder;
 import org.talend.daikon.model.GitCommit;
 import org.talend.daikon.model.MiscReleaseNoteItem;
 import org.talend.daikon.model.ReleaseNoteItem;
