+++ releases/1.9.0.adoc
@@ -1,4 +1,4 @@
-= Daikon Release Notes (1.9.0)
+= Daikon Release Notes (1.9.0) - 11/05/2019
 
 Thanks to @Henri-Mayeul de Benque, @dependabot[bot], @Michael Tallet, @tduperron, @Francois Huaulme, @iwangtalend
